change doubles to floats

+++ lib/lib_i2c/SPL06_007/src/SPL06-007.cpp
@@ -54,22 +54,22 @@ uint8_t get_spl_fifo_sts()
 	return i2c_eeprom_read_uint8_t(SPL_CHIP_ADDRESS, 0x0B);
 }
 
-double get_traw_sc() {
+float get_traw_sc() {
 	int32_t traw = get_traw();
-	return (double(traw)/get_temperature_scale_factor());
+	return (float(traw)/get_temperature_scale_factor());
 }
 
-double get_temp_c() {
+float get_temp_c() {
 	int16_t c0,c1;
 	c0 = get_c0();
 	c1 = get_c1();
-	double traw_sc = get_traw_sc();
-	return (double(c0) * 0.5f) + (double(c1) * traw_sc);
+	float traw_sc = get_traw_sc();
+	return (float(c0) * 0.5f) + (float(c1) * traw_sc);
 }
 
-double get_temperature_scale_factor()
+float get_temperature_scale_factor()
 {
-  double k;
+  float k;
   uint8_t tmp_Byte;
   tmp_Byte = i2c_eeprom_read_uint8_t(SPL_CHIP_ADDRESS, 0X07); // MSB
   //tmp_Byte = tmp_Byte >> 4; //Focus on bits 6-4
@@ -78,35 +78,35 @@ double get_temperature_scale_factor()
   switch (tmp_Byte) 
   {
     case 0B000:
-      k = 524288.0d;
+      k = 524288.0f;
     break;
 
     case 0B001:
-      k = 1572864.0d;
+      k = 1572864.0f;
     break;
 
     case 0B010:
-      k = 3670016.0d;
+      k = 3670016.0f;
     break;
 
     case 0B011:
-      k = 7864320.0d;
+      k = 7864320.0f;
     break;
 
     case 0B100:
-      k = 253952.0d;
+      k = 253952.0f;
     break;
 
     case 0B101:
-      k = 516096.0d;
+      k = 516096.0f;
     break;
 
     case 0B110:
-      k = 1040384.0d;
+      k = 1040384.0f;
     break;
 
     case 0B111:
-      k = 2088960.0d;
+      k = 2088960.0f;
     break;
   }
   return k;
@@ -131,13 +131,13 @@ int32_t get_traw()
   return tmp;
 }
 
-double get_praw_sc()
+float get_praw_sc()
 {
 	int32_t praw = get_praw();
-	return (double(praw)/get_pressure_scale_factor());
+	return (float(praw)/get_pressure_scale_factor());
 }
 
-double get_pcomp()
+float get_pcomp()
 {
 	int32_t c00,c10;
 	int16_t c01,c11,c20,c21,c30;
@@ -148,19 +148,19 @@ double get_pcomp()
 	c20 = get_c20();
 	c21 = get_c21();
 	c30 = get_c30();
-	double traw_sc = get_traw_sc();
-	double praw_sc = get_praw_sc();
-	return double(c00) + praw_sc * (double(c10) + praw_sc * (double(c20) + praw_sc * double(c30))) + traw_sc * double(c01) + traw_sc * praw_sc * ( double(c11) + praw_sc * double(c21));
+	float traw_sc = get_traw_sc();
+	float praw_sc = get_praw_sc();
+	return float(c00) + praw_sc * (float(c10) + praw_sc * (float(c20) + praw_sc * float(c30))) + traw_sc * float(c01) + traw_sc * praw_sc * ( float(c11) + praw_sc * float(c21));
 }
 
-double get_pressure()
+float get_pressure()
 {
 	return get_pcomp() / 100; // convert to mb
 }
 
-double get_pressure_scale_factor()
+float get_pressure_scale_factor()
 {
-	double k;
+	float k;
 
 	uint8_t tmp_Byte;
 	tmp_Byte = i2c_eeprom_read_uint8_t(SPL_CHIP_ADDRESS, 0X06); // MSB
@@ -171,35 +171,35 @@ double get_pressure_scale_factor()
 	switch (tmp_Byte) // oversampling rate
 	{
 		case 0B000:
-			k = 524288.0d;
+			k = 524288.0f;
 		break;
 
 		case 0B001:
-			k = 1572864.0d;
+			k = 1572864.0f;
 		break;
 
 		case 0B010:
-			k = 3670016.0d;
+			k = 3670016.0f;
 		break;
 
 		case 0B011:
-			k = 7864320.0d;
+			k = 7864320.0f;
 		break;
 
 		case 0B100:
-			k = 253952.0d;
+			k = 253952.0f;
 		break;
 
 		case 0B101:
-			k = 516096.0d;
+			k = 516096.0f;
 		break;
 
 		case 0B110:
-			k = 1040384.0d;
+			k = 1040384.0f;
 		break;
 
 		case 0B111:
-			k = 2088960.0d;
+			k = 2088960.0f;
 		break;
 	}
